consider expired leases rotatable
this helps mitigate secret expiration (outside the control
of the client such as in a network outage) that would otherwise
be rotatable. before this change, a lease wasn't marked
'renewable false' during the run of manage-leases so it
would just be swept. instead, this at least attempts to
rotate first before sweeping it for being expired