fix tests to match reality, be internally consistent handling non-2XX responses
if a function or caller still wants to handle
404, supports-not-found is the recommendation.

this was an unfortunate case of getting overly detailed
in setting up unit test state, where the actual response
from the http client lib was mocked and thus was insulated
from change since the mock behavior didn't match up
with actual library behavior changes.

this should mostly be replaced with integration tests
and the subroutines and other functions be tested as
black box input/output data.